The need for MobileTwigModule has decreased, so we will make it an optional install.
See https://github.com/madapaja/Madapaja.TwigModule/releases/tag/2.6.0

In order to use the MobileTwigModule, it is necessary to install mobiledetect/mobiledetectlib. Relocate `mobiledetect/mobiledetectlib` to "require-dev" in composer.json and implement a runtime check that throws an exception if the dependency is missing in MobileTwigModule. Additionally, add getInput method to ErrorPagerRenderer to handle input values from resource objects.